Remove experimentalRefreshOnResume
This never actually worked, the problem is that the terminal is still hanging
onto the old texture instead of disposing it which was the intent. Some way
to indicate to re-create the texture atlas would be needed to fix this, but
the hope is to remove the canvas renderer eventually anyway.

Part of #69665