suggest: fix handling around long lines
See: https://github.com/microsoft/vscode/issues/90552#issuecomment-592680419

`overflow: hidden` should have been `overflow: auto`. Auto displays a
scrollbar which, depending on the platform, can make the items too high.

Also fixes names being cut off (see the loooong) cutoff in the linked
issue. `flex-shrink:0` is on the label, with a max-width of 100%. But
on the left there was the 18px icon, so the right side of the label was
18px off the end and not visible. Fix it by moving the icon outside of
the `.left` side. We could alternately `calc(100% - 18px)`, but since
the icon was not a hardcoded size in CSS I didn't want to implicitly
depend on that.