Restore old remote uri normalization for remote uris
For #100442

When we moved the webview resource protocol to the main process, I updated the uri normalization logic so that it could load remote resources properly. This however broke VSO since their resources are still loaded from the renderer process.

This change restores the old normalization for the renderer side resource loader.