Make sure getEditorOverrides can return explicitly requested editors
Fixes #100238

For cases where an extension explicitly requests that a given editor be used, such as `openWith`, `getEditorOverrides` would not always include the requested editor in its returned values.

This change fixes that by making `getEditorOverrides` take an optional `id`. Providers can look at this id to determine if they should return a specific editor override even it it does not match the resource