Add delegation back to renderer process for cetain webview resource reeds (#102442)
Fixes #102191

**Problem**
Webviews used to load local resources in the renderer process. This gave them access to all of the file system providers registered for that renderer (such as `untitled` and `git`).

However, last iteration we moved the webview protocol to the main process. This works fine for file resources, but the main process cannot read fancy uris such as `git:` or `untitled:`

**Fix**
To fix this, I've added code that delegates resource requests for non-file uris from the main process back to the renderer process. We tried to avoid this as it adds another round trip to each request, but I don't see a better way of doing this.

This problem only effects desktop VS Code. Web VS Code uses a service worker which always reads resources through the renderer process