Pass webview editors an explicit scoped context key service
Fixes #100953

Because webviews are not part of the normal dom, they also are not part of the normal editor context structure.

This change fixes that by allowing a webview to be passed an scoped context key service to use. Inside the webview, we then create a scoped context key service that flows back into the normal editor dom