trees: add diffIdentityProvider for efficient setChildren updates
This uses the existing `LcsDiff` diff implementation to generate
diff operations when calling `setChildren` on a tree view where the new
`diffIdentityProvider` is given. I didn't make this 'default' using
the existing identityProvider, since I noticed that (at least) the
Explorer view injects different elements in `setChildren` that have the
same identity. Previously these would have been replaced, but now they
wouldn't. This may be an effect of compressed trees.

https://github.com/microsoft/vscode/blob/f897ca87c88e632dc787b66da6ed2d72b466b069/src/vs/base/browser/ui/tree/compressedObjectTreeModel.ts#L151-L162

Speaking of compressed trees -- this doesn't work there yet. It looks
like compressed trees call setChildren with the parent of the element
for whom children are being set... this worked before since the update
replaced all previous children, but now the update is not deep. I
haven't dug into this yet to find out why this is needed.

Performance numbers. This is much faster for the large test tree.
Previously each update, by the time all tests were discovered, took
about 150ms per tree update:

![](https://memes.peet.io/img/21-01-9cbf70b4-6752-4d0a-82b7-f26bd02db01d.png)

Now it's an average of 16.3ms per call, and doesn't take longer for
larger trees. Noticably faster to finish and not laggy while it does.

![](https://memes.peet.io/img/21-01-a8d757b8-0529-4067-9d8c-45138d9db7bc.png)

We should consider adopting this for all trees if/when we deal with
the changed setChildren semantics.