Make sure we always apply TS auto imports, even if VS Code applies the completion before it has been resolved
Fixes #109439

This introduces a new `ApplyCompletionCommand` that is included on all JS/TS completions, which  applies additional parts of the completion (such as auto imports).

This is needed since VS Code will not always wait until `resolveCompletionItem` completes before appling the completion. This causes auto imports to sometimes not work when typing quickly