Fix remote terminal reconnect
Remote terminals were being saved on the server using the 'persistent
terminal ID' which is correct, but on the client side they were using
the 'terminal instance ID'. This works fine until a terminal gets relaunched
or a second window is opened, at which point the 2 IDs no longer align with
each other so restoring becomes flaky (some terminals may restore but it's
mostly broken).

This change uses the persistent terminal ID correctly in the remote's impl
of ITerminalChildProcess and falls back to 0 if it's not available (to be
filtered out all together in microsoft/vscode#117971). An extra event
listener for ready was also added to trigger the remote layout update since
that fires when persistentTerminalId is ready to go.

Fixes #117896