Fix task start firing with invalid process ID in dependsOn task
This was caused by the env var relaunch @debounce decorator that was added to
prevent the process from relaunching multiple times when multiple extensions
contribute to the environment. This caused the reuseTerminal call to not have
the new processReady promise ready immediately after reuseTerminal was called.
The fix which seems safe it to just move the @debounce over to relaunchTerminal
that tasks doesn't use.

Fixes #118256