perf - change lifecycle phase restored to work differently
Previously the restored and eventually phase would only be reached when all visible editors have loaded.
This means, a slow editor will block a lot of useful workbench contributions from working.
A better approach is to give a bit of time to the visible editors to restore, but then signal the restored and eventually phase anyway.