Make sure we layout editors in sideBySideEditor after they are created (#125445)
Fixes #125202

This make sure we always invoke `layout` when a new editor is created in a split view. Due to the caching added in 524d856299fadf9e0e708dbb8880f34593ab9b30, the `WebviewEditor` created by the split view never had `layout` called on it

Not sure if this is best fix for this issue. If not, I suggest we revert 524d856299fadf9e0e708dbb8880f34593ab9b30 and then revisit it next iteration