Re-enable webview tests on web (#134300)
* Enable webview tests on web

These tests pass fine on Chrome and webkit. On Firefox, the two about loading resources currently fail

I spent a while trying to debug what is going on with these tests. It seems like the service workers is registered properly but `fetch` is never called on it. Skipping these for now

* Remove only

* Fix test