Updating the TerminalInstance to only reigster link providers once. (#135419)
When a task reuses a terminal, the TerminalProcessManager would fire the processReady event each time a task is run, resulting in multiple registrations of the link providers. Re-registering the terminal link providers would cause xterm to trigger multiple overlapping providers on the same line of output, which is not supported by xterm at the moment. The net result is that no external link providers would work for terminals that are reused by tasks.