glyphMargin: Wrapped codicons inside parent div
This patch resolves #137436.
When multiple classes (e.g. codicon-debug-breakpoint,codicon-debug-hint)
are added they would disable the ::after content for consecutive classes
for the HTML Element.

This patch adds a parent element to the overlay and adds all classes as
seperated divs, so they can be drawn on top of each other.