Restrict vscode-file requests to main window (#139140)
* Restrict vscode-file requests to main window

This change tries to limit use of the vscode-file protocol to the main window.

While investigating this issue, I hit a bug: `CodeWindow` registers a `onBeforeRequestHandler` but can then be disposed of. When this happens, the handler remains active but any references to `this` inside of it now reference the disposed class.

To fix this, I've moved all the request filtering logic into `app.ts` instead and register it on the `defaultSession` instead of on a web contents.

Then to fix the original issue, I added logic that checks the routingId of the request against the routingIds of all known windows. Only windows with known routingIds are allowed to use `vscode-file`.

* Move logic into configure session

* Check processId

* Check BrowserWindow.getAllWindows instead of just main VS Code windows