Fix editor eating ContextMenu key when contexmenu is disabled
The `EditorOption.contextmenu` setting is supposed to disable the context menu so users can provide their own implementation. However, it is eating the context menu keydown event so the user-defined context menu can only be activated with the mouse and not the keyboard.

This fixes the issue by checking the setting when the keydown event fires.