Scope drop into editor to specific editors (#147425)
* Scope drop into editor to specific editors

As part of #142990, I added support for dropping dragged files from the desktop into an editor by holding shift. To help users discover this feature, we also show a prompt in the editor overlay when dragging

However there are a number of editor types that don't actually support drops. The welcome view for example. Readonly editors should also not support dropping. In those cases, it is confusing the show the `hold shift to drop into editor` prompt

This change adds a new editor capability (`CanDropIntoEditor`) which signals that the editor does support dropping into the editor. I then enable this new capability on writable files, webviews, and notebooks

* Making suggested changes