fix: compare origins in fetch requests against remote authority (#147467)
The previous implementation did not consider what would happen if
webview resources were served from the same domain. By first comparing
the requestUrl.orgin with the sw.orgin (similar to how it's done for
localhost), this is no longer a problem.

And since the requests have the same origin, authentication will never
be an issue as cookies will exist.