Always show overview ruler decorations before command execution
Originally we hid the overview ruler decoration for command currently being input
and executed. This ended up causing problems for commands that took a while to run
and task decoration support as they would not get a mark in the overview ruler
until the command finished. For simplicity let's just always show this, so there
will basically always be a grey mark at the very bottom for the cursor line, but
that is at least consistent and then we don't need to juggle different overview
ruler options depending on the state of the command.

Fixes #152663