remember view containers that aren't registered yet (#153150)
fixes #149503

The comments panel was being registered later so by the time we went to
save it back to the cache, we thought were removing it. Then, once it
was registered, we would pin it since we thought it was new. Following,
the behavior of the activity bar, we save it back to the cache as is.