shellIntegrationAddon: fix incorrect deserializeMessage implementation
The original implementation confused escaped and un-escaped sequences
due to squashing each adjacent pair of backslashes prior to parsing.

Unlike encoding, decoding cannot be performed in passes this way: it
must be sequential, because of the potential for overlapping patterns
(e.g. the third backslash in "\\\x3b").

This replaces the implementation with a single sequential regex, which
matches the escape sequences and replaces each match. This enables the
tests that were broken in the previous implementation.

To illustrate, given the following two different original values (here
as literals between «», without any escaping):
  a. «Packing\Stuff\x3BEarmuffs»
  b. «Packing\Stuff;Earmuffs»

Those should be distinctly encoded as:
  a. «Packing\\Stuff\\x3BEarmuffs»
  b. «Packing\\Stuff\x3BEarmuffs»

The original implementation wrongly threw away the escaping information
by replacing each adjacent pairs of backslashes with a single backslash,
regardless of whether escaping was in effect or not:
  a. «Packing\Stuff\x3BEarmuffs»
  b. «Packing\Stuff\x3BEarmuffs»

The new implementation matches, in (a), each "\\"; and there are no
non-overlapping "\x…" sequences. In (b), the first "\\" is matches, and
the "\x3B" is matched.

This works for any correct combination of adjacent escape squences.