shellIntegrationAddon: fix confusion of 'no match' and 'match at start of string'
deserializeMessage() would not recognize escape sequences at the start
of a string, due to a simple falsey check on `match?.index` which could
be validly 0 instead of null.

In other words: "Foo\x3bBar" would decode to "Foo;Bar", but "\x3b" would
not be recognized as an escape sequence.

This fixes that bug, enabling those suppressed test cases, and disables
the newly-failing tests which had only passed due to the (still-broken)
escaping logic being skipped in those situations.