Lazily compute icon classes of quick input items (#163886)
Lazily compute icon classes on quick input items

When showing a file list in quick input, we are eagerly resolving all the icons (even for items that are not visible). In the vscode workspace, this results in 512 calls into `getLanguageIds` every time the full quick input is updated (512 is the max number of quick input files we show)

This updates the icon classes to instead be resolved lazily when the quick input items are rendered. This cuts the cost of `getFilePicks` in half