Use `IAction` instead of `Action` for search item action (#164844)
For actions that are static, it is better to use `IAction` instead of `Action`. This avoids creating an extra disposable, emitter, and event listener (when used with an action bar). It also means you don't have to dispose of the action

This small optimization is useful here because we create one of these actions for every search item in the list as you scroll