Allow editor contributions to mark themselves as instantiated on idle (#166065)
* Allow editor contributions to mark themselves as instantiated on idle

For #164171

Editor contributions are currently created eagerly when editors are instantiated. In many cases however, the contribution is not needed instantly. A contributed widget for instance may only be shown after a user action, while some info that shows in the editor also does not need to be shown instantly as soon as the editor opens

Contributions like those can be lazily created, but actually adopting lazy instantiation for individual contributions is a lot of work

Instead this PR lets editor contributions tell VS Code that the entire contribution does not need to be eagerly created and can be instead on idle

With `EditorContributionInstantiation.Idle`, the contribution will be created:

- Either when the editor is idle
- Or when the contribution is explicitly requested.

Idle contributions cannot participate in saving and restoring of editor view states

This change will make it easier to lazily create editor contributions and also lets us remove existing code that was added to make individual editor contributions lazy

* Mark as optional

* Addressing comments

* IdleValue should throw again if there's ctor error