Don't create more than one notebook detection task per-type (#167878)
Don't create more than one notebook detection task per-type.
This is creating detection tasks to cover the period between opening a notebook, and the notebook extension activating and starting its own detection task. This code is a bit more complex than necessary, and it would be simpler to manage the lifecycle of this task from the `activateByEvent` promise.
Fix #167875