expose fish integration configuration via `XDG_DATA_DIRS`
closes #139400

`XDG_DATA_DIRS` is specified[1] as
> a set of preference ordered base directories relative to which data files should be searched.

the fish shell uses it[2] (among other things)
to let *other tools* provide additional configuration files
for the fish shell to load on startup.

this change exposes the fish shell integration configuration script
in a place matching what fish expects
(in a `fish/vendor_conf.d` subdirectory)
and appends the exposed directory to the value of `XDG_DATA_DIRS`
if automatic shell integration is requested.
if `XDG_DATA_DIRS` is unset it uses the specified default value
`/usr/local/share:/usr/share`
and appends the exposed xdg_data directory to that.

[1]: https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html
[2]: https://fishshell.com/docs/current/language.html#configuration-files